refactor sparql results container
