--- src/subapps/studioLegacy/containers/DashboardResults/DashboardElasticSearchQueryContainer copy.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-const DashboardElasticSearchQueryContainer: React.FC = () => {};
-
-export default DashboardElasticSearchQueryContainer;
