+++ src/subapps/studioLegacy/containers/DashboardResults/DashboardElasticSearchQueryContainer.tsx
@@ -1,0 +1,13 @@
+import { ElasticSearchView } from '@bbp/nexus-sdk';
+import * as React from 'react';
+
+const DashboardElasticSearchQueryContainer: React.FC<{
+  view: ElasticSearchView;
+  dataQuery: string;
+  dashboardLabel: string;
+  goToStudioResource: (selfUrl: string) => void;
+}> = ({ view, dataQuery, dashboardLabel, goToStudioResource }) => {
+  return <>"banana"</>;
+};
+
+export default DashboardElasticSearchQueryContainer;
