+++ src/subapps/studioLegacy/containers/DashboardResultsContainer.tsx
@@ -1,107 +1,14 @@
 import * as React from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
-import { Spin, Alert, message } from 'antd';
-import {
-  SelectQueryResponse,
-  SparqlViewQueryResponse,
-  Resource,
-  View,
-  NexusClient,
-} from '@bbp/nexus-sdk';
+import { message, Skeleton } from 'antd';
+import { ElasticSearchView, Resource, SparqlView, View } from '@bbp/nexus-sdk';
 import { useNexusContext } from '@bbp/react-nexus';
+import { match, when } from 'ts-pattern';
 
-import ResultsTable, {
-  ResultTableProps,
-} from '../../../shared/components/ResultsTable/ResultsTable';
-import { camelCaseToLabelString, parseProjectUrl } from '../../../shared/utils';
+import { parseProjectUrl } from '../../../shared/utils';
 import useAsyncCall from '../../../shared/hooks/useAsynCall';
-
-export interface QuerySparqlViewWithDataQueryProps {
-  nexus: NexusClient;
-  orgLabel: string;
-  projectLabel: string;
-  dataQuery: string;
-  viewId: string;
-}
-
-const querySparqlViewWithDataQuery = ({
-  nexus,
-  orgLabel,
-  projectLabel,
-  dataQuery,
-  viewId,
-}: QuerySparqlViewWithDataQueryProps) => async () => {
-  const view = await nexus.View.get(orgLabel, projectLabel, viewId);
-  if (view && view['@type']?.includes('ElasticSearchView')) {
-    const result = await nexus.View.elasticSearchQuery(
-      orgLabel,
-      projectLabel,
-      encodeURIComponent(viewId),
-      dataQuery
-    );
-    return result;
-  }
-  const result: SparqlViewQueryResponse = await nexus.View.sparqlQuery(
-    orgLabel,
-    projectLabel,
-    encodeURIComponent(viewId),
-    dataQuery
-  );
-  const data: SelectQueryResponse = result as SelectQueryResponse;
-  const tempHeaderProperties: {
-    title: string;
-    dataIndex: string;
-  }[] = data.head.vars
-    .filter(
-      // we don't want to display total or self url in result table
-      (headVar: string) => !(headVar === 'total' || headVar === 'self')
-    )
-    .map((headVar: string) => ({
-      title: camelCaseToLabelString(headVar), // TODO: get the matching title from somewhere?
-      dataIndex: headVar,
-    }));
-  const headerProperties = tempHeaderProperties;
-  // build items
-  const items = data.results.bindings
-    // we only want resources
-    .filter((binding: Binding) => binding.self)
-    .map((binding: Binding, index: number) => {
-      // let's get the value for each headerProperties
-      const properties = tempHeaderProperties.reduce(
-        (prev, curr) => ({
-          ...prev,
-          [curr.dataIndex]:
-            (binding[curr.dataIndex] && binding[curr.dataIndex].value) ||
-            undefined,
-        }),
-        {}
-      );
-      // return item data
-
-      return {
-        ...properties, // our properties
-        id: index.toString(), // id is used by antd component
-        self: binding.self, // used in order to load details or resource once selected
-        key: index.toString(), // used by react component (unique key)
-      };
-    });
-  return {
-    headerProperties,
-    items,
-  };
-};
-
-export type Binding = {
-  [key: string]: {
-    dataType?: string;
-    type: string;
-    value: string;
-  };
-};
-
-type NexusSparqlError = {
-  reason: string;
-};
+import DashboardSparqlQueryContainer from './DashboardResults/DashboardSparqlQueryContainer';
+import DashboardElasticSearchQueryContainer from './DashboardResults/DashboardElasticSearchQueryContainer';
 
 const DashboardResultsContainer: React.FunctionComponent<{
   dataQuery: string;
@@ -139,44 +46,39 @@ const DashboardResultsContainer: React.FunctionComponent<{
     [orgLabel, projectLabel, viewId]
   );
 
-  const queryResult = useAsyncCall<
-    {
-      headerProperties: ResultTableProps['headerProperties'];
-      items: ResultTableProps['items'];
-    },
-    Error | NexusSparqlError
-  >(
-    querySparqlViewWithDataQuery({
-      nexus,
-      orgLabel,
-      projectLabel,
-      dataQuery,
-      viewId,
-    })(),
-    [dataQuery, viewId]
-  );
-
-  if (queryResult.error) {
-    return (
-      <Alert
-        message="Error loading dashboard"
-        description={`Something went wrong. ${(queryResult.error as NexusSparqlError)
-          .reason || (queryResult.error as Error).message}`}
-        type="error"
-      />
-    );
-  }
-
-  return (
-    <Spin spinning={queryResult.loading}>
-      <ResultsTable
-        headerProperties={queryResult.data?.headerProperties}
-        items={queryResult.data?.items || []}
-        handleClick={goToStudioResource}
-        tableLabel={dashboardLabel}
-      />
-    </Spin>
-  );
+  return match(viewResult)
+    .with({ loading: true }, () => <Skeleton active={true} />)
+    .with(
+      {
+        error: null,
+        data: when(data => !!(data && data['@type']?.includes('SparqlView'))),
+      },
+      () => (
+        <DashboardSparqlQueryContainer
+          view={viewResult.data as SparqlView}
+          dataQuery={dataQuery}
+          dashboardLabel={dashboardLabel}
+          goToStudioResource={goToStudioResource}
+        />
+      )
+    )
+    .with(
+      {
+        error: null,
+        data: when(
+          data => !!(data && data['@type']?.includes('ElasticSearchView'))
+        ),
+      },
+      () => (
+        <DashboardElasticSearchQueryContainer
+          view={viewResult.data as ElasticSearchView}
+          dataQuery={dataQuery}
+          dashboardLabel={dashboardLabel}
+          goToStudioResource={goToStudioResource}
+        />
+      )
+    )
+    .run();
 };
 
 export default DashboardResultsContainer;
